# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "User API's",
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "User Registration",
                "description": "Register by ...",
                "operationId": "registerUser",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "email",
                                    "password",
                                    "phone",
                                    "code_phrase"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Username"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "email@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "992999999999"
                                    },
                                    "code_phrase": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "ih"
                                    },
                                    "devices": {
                                        "description": "Devices of user form Devices table",
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "format": "id"
                                        },
                                        "collectionFormat": "multi",
                                        "example": "[1,2]"
                                    },
                                    "status": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "409": {
                        "description": "User already exist",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Duplicate (phone or email) user on register."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "Correct Registration",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9.eyJhdWQiOiIxIiwianRpIjoiM2Q5MTkxYmE3OWNmYzM3MzcyMWIzNmFlOWJlMTRkMTljYjc2YWRhZGQ0NDA3MzIxNzczMzRkYmY1OTc3MWUwODc0OWNmZWJkNjBjNThjZjQiLCJpYXQiOjE3MDU2ODc2OTcuOTAwMDg3LCJuYmYiOjE3MDU2ODc2OTcuOTAwMDksImV4cCI6MTczNzMxMDA5Ny44ODY2NjEsInN1YiI6IjQiLCJzY29wZXMiOltdfQ.Qrf05e_exPvESGJIeSuOpF56kH7drJsd9L3ZymFkyi68ZOoCFvqesRU505p10JUwNQ14dyrlOPpjq85sk6AdEVHT0vaO2WRspaHVrF3ZPSyNH3VpuVRfs12lGNlFaOMTzR1O1qb2PLL306KrltPfwDa1G42GH6Sl_ji2Aiu-2-Rsc0ap6dgFSI5GmiwO92ErE_nMgMjfNwvmw3CgPPWhs934WcZaBHjAYU5csbYoNxq8SsQ5IxCcWyhoj4Tm2D-hbhEtOccYAmG71Cen_z8DOf2Zso20eDKWBZdNynWvpKXyHJYh1wyB2aBQKcHqDMa-dvoQg5_i1-Z7AELG-C1GkqiTo7qPwT5npJz8kpmIhlM9vkh3HelWHajkcbWDw1HZV--wFZJEdAuBlA7j9YffFsVmF1rmyxbA6fNGKCpqaLYEqDXNo_nMkI9NcP_Rwcd5TFSACXw17xqPVzZndmyar06r2ytYHSXPmUorq7xzYVYg1087KS2y6ZQQ2KUkcQlL4qKzHJS2gU-xQUe3gQEzN2cb6-PBbM-w99o2qaH-z4oR8NTY481NXGfEQFLPAxwGkONak1JYVewlqztltenqT2wfzcYwVaokQ7CIbgKHRuhyyWkkFqf9q1Lnkc3CwtoMmSb7RXs5bq2Ia3obGx6cKOgOpvGlM6zo0eC7FDsFa5s"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Name"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "User Authorization",
                "description": "Auth by ...",
                "operationId": "authUser",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "phone": {
                                        "type": "string",
                                        "example": "992999999999"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "User not found on auth."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Password is incorrect",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Password is incorrect."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Signed in",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9.eyJhdWQiOiIxIiwianRpIjoiM2Q5MTkxYmE3OWNmYzM3MzcyMWIzNmFlOWJlMTRkMTljYjc2YWRhZGQ0NDA3MzIxNzczMzRkYmY1OTc3MWUwODc0OWNmZWJkNjBjNThjZjQiLCJpYXQiOjE3MDU2ODc2OTcuOTAwMDg3LCJuYmYiOjE3MDU2ODc2OTcuOTAwMDksImV4cCI6MTczNzMxMDA5Ny44ODY2NjEsInN1YiI6IjQiLCJzY29wZXMiOltdfQ.Qrf05e_exPvESGJIeSuOpF56kH7drJsd9L3ZymFkyi68ZOoCFvqesRU505p10JUwNQ14dyrlOPpjq85sk6AdEVHT0vaO2WRspaHVrF3ZPSyNH3VpuVRfs12lGNlFaOMTzR1O1qb2PLL306KrltPfwDa1G42GH6Sl_ji2Aiu-2-Rsc0ap6dgFSI5GmiwO92ErE_nMgMjfNwvmw3CgPPWhs934WcZaBHjAYU5csbYoNxq8SsQ5IxCcWyhoj4Tm2D-hbhEtOccYAmG71Cen_z8DOf2Zso20eDKWBZdNynWvpKXyHJYh1wyB2aBQKcHqDMa-dvoQg5_i1-Z7AELG-C1GkqiTo7qPwT5npJz8kpmIhlM9vkh3HelWHajkcbWDw1HZV--wFZJEdAuBlA7j9YffFsVmF1rmyxbA6fNGKCpqaLYEqDXNo_nMkI9NcP_Rwcd5TFSACXw17xqPVzZndmyar06r2ytYHSXPmUorq7xzYVYg1087KS2y6ZQQ2KUkcQlL4qKzHJS2gU-xQUe3gQEzN2cb6-PBbM-w99o2qaH-z4oR8NTY481NXGfEQFLPAxwGkONak1JYVewlqztltenqT2wfzcYwVaokQ7CIbgKHRuhyyWkkFqf9q1Lnkc3CwtoMmSb7RXs5bq2Ia3obGx6cKOgOpvGlM6zo0eC7FDsFa5s"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/user/{id}": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Get User",
                "description": "Get User by Id",
                "operationId": "getUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "User identifier",
                        "required": true,
                        "example": "12"
                    }
                ],
                "responses": {
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "User not found on auth."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "User founded",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "12"
                                        },
                                        "username": {
                                            "type": "string",
                                            "example": "Abdurazzoq"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "abdurazzoq9050@gmail.com"
                                        },
                                        "phone": {
                                            "type": "string",
                                            "example": "992928369050"
                                        },
                                        "code_phrase": {
                                            "description": "Code Phrases",
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "format": "id"
                                            },
                                            "example": "['ih', 'liebe', 'dich']"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "active"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2024-01-21T09:09:38.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2024-01-21T09:09:38.000000Z"
                                        },
                                        "devices": {
                                            "description": "List of devices",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "1"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Poco M3"
                                                    },
                                                    "ip": {
                                                        "type": "string",
                                                        "example": "192.168.1.100"
                                                    },
                                                    "location": {
                                                        "type": "string",
                                                        "example": null
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2024-01-21T08:39:46.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2024-01-21T08:39:46.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}